Fix alpha double blend
The color when used alpha factor always got darker. The problem was a double blend that produced a reduction in the color saturation.

Now, there is a premult of the alpha factor before doing last blend pass.

As a result of this change, now the drawing process can be simplified and a pass can be removed. Some cleanup done too.

Still need more work in render mode. Maybe the problem in render is related to Tonemapping.

Thanks Clément Foucault for his help solving this issue.