Initial working drag & drop for workspace tabs
Had to do quite some changes to get this to work, mainly because we'd
want to have the customized order written to files. Also wanted to use
existing wmDrag and wmDropBox code but modify it for our needs.
And of course, this should not just work for a single case, code should
be applicable to other areas as well.

Note that file read/write is not working (yet), neither is versioning, region
copying and deleting workspaces working.
Code is also very WIP, much of it is placed in temporary places and
such.

Once the basics are done, we can add nicer feeback during drag & drop.