Move data of global areas into a separate struct
Made the struct accessible via a pointer in the ScrArea so that we can
use a simple NULL-check for this pointer to see if the area is global.

The struct is not written to files, meaning we have the freedom to do
changes to it without having to worry about compatibility.