WIP: Quick hack to get Timeline header buttons to show up where we want them to
WARNING: Prototype code to test that the UI design idea works.

This is a nasty hack, as it uses hardcoded button sizes (extracted by hand from
a screenshot and tweaked), and used to adjust the scale_x of layout items used
to house the spacers-layouts.

(The non-hacky part of this the new "uiItemSpacer" widget. Eventually the layout
manager should be identifying these widgets and setting the sizes on them in using
a method similar to what we're doing here).