Fix: Motionpaths were not drawing for bones
As with the objects case earlier, the problem was that when doing COW for the
viewport, the pchan->mpath pointer was just getting set to NULL in the copying
code. It's unclear for now whether this will cause any problems down the track
somewhere else - the comment there was non-specific about why we were explicitly
not copying/keeping the old link.