Add a new User Prefs parameter to control GP multisample AA
For general use of Viewport the multisample AA is not required and can produce errors in selection, but for grease pencil the AA is required all the time to get smooth lines, so a new parameter has been added to User Prefs screen to define this value.

If we use the same value for both AA (grease pencil and Viewport), the viewport would be intefered by grease pencil settings.

The new value is below Viewport Samples.

By default the value was set to 4 samples.

This changes was done as result of requirement of grease pencil development team artists.