Add first parts of UDIM support in the core image code
Internally, UDIMs are handled very similar to sequences - they consist of a list
of files (containing a sequential number in their name) that all belong to one
Image datablock. This means that existing code like the IBuf cache can be reused.

The current drawing and opening code are just a minimal implementation to test
the core code, they'll be improved of course.