Eevee: LightCache: Write LightCache to File
This patch does a few things for it to work.
- It port some eevee's struct to DNA for them to be saved in the file.
- Readback the result of the lightcache baking and store it in the scene
  struct.

For small scenes, it's okayish to keep the data around in memory but a
better approach would be to save it to disk.

A standard reflection probe is more than 6 MB of data.
Diffuse irradiance grids are relatively small compared to this.