Annotations: Restore 3D View support for annotations  (WIP)
Initial support for getting annotations to show up in the 3D
viewport again.  WIP and slightly hacky, but it restores basic
working functionality so we can focus on polishing the "new" parts.

Ultimately, the way these callbacks are integrated into the
drawing loop will probably need to change. For example,
* Perhaps we might need a full overlay engine for this one case
  (as for motionpaths), though that may also be overkill, and would
  further split the code in different places.
* The draw order (relative to manipulators) needs review. It probably
  should be below both sets of manipulators, not sandwiched between them.

Other notes:
* For this to work, we currently need to use the original (non-evaluated)
  scene data, or else we don't get the newly drawn strokes, as the
  scene->gpd stuff isn't copied by COW yet.