Icons for Grease Pencil/Annotation tools
These were in the icon_geom.blend file in SVN (lib/resources),
but needed to be manually added to the Export group for this to
work.

They are needed for use for use with the annotation tools.

NOTE: There are currently problems running the icon generation
scripts in the gp-object branch due to some file loading pointer
remapping problems there causing segfaults when running the script.
Thus, I generated these icons using the 2.8 branch, then cherry-picked
that commit over to this branch.