try to avoid crashes due to excessive usage of immediate draw buffer
since the buffersize is limited to 4 MB (but it effectively crashed above usage of 3 MB,
probably due to 4 bit padding), there is a hard limit of usage ratio to be taken into account. The debug drawer counts the bytes used and increases an epsilon value as softlimit
first. This filters out short lines / very close verts and reduces the buffer usage. If that is not enough, there is a hard limit at 95% buffer usage. Then line drawing is skipped.