GP: Mew Merge Strokes operator
This operator allows to create a new stroke joining several stroke selected points of different strokes.

The new stroke will use the current material.

To use, first select the points to be merged. Optionally can remove the old points and strokes.

The operator is available in Edit mode in the Specials menu and Stroke menu.