GP: Set Cap mode as Enums
Now the Cap mode is an enum ready for future modes.

Add the cap mode to material is not logic because the stroke start and end depends on the drawing direction. Maybe in the future we could add to material, but will need more work at stroke level.

The Toggle operator keeps equal in a circular switch mode. We can improve it also in the future.